TOMEE-280 better management of jars + include/exclude property
git-svn-id: https://svn.apache.org/repos/asf/openejb/trunk/openejb@1361124 13f79535-47bb-0310-9956-ffa450edef68

+++ container/openejb-core/src/main/java/org/apache/openejb/assembler/classic/DeployTimeEnhancer.java
@@ -1,20 +1,14 @@
 package org.apache.openejb.assembler.classic;
 
 import org.apache.openejb.config.event.BeforeDeploymentEvent;
-import org.apache.openejb.config.sys.ListAdapter;
-import org.apache.openejb.config.sys.ServiceProvider;
 import org.apache.openejb.loader.Files;
+import org.apache.openejb.loader.SystemInstance;
 import org.apache.openejb.observer.Observes;
-import org.apache.openejb.util.JarExtractor;
-import org.apache.openejb.util.LogCategory;
-import org.apache.openejb.util.Logger;
-import org.apache.openejb.util.URLs;
+import org.apache.openejb.util.*;
 import org.xml.sax.Attributes;
-import org.xml.sax.HandlerBase;
 import org.xml.sax.SAXException;
 import org.xml.sax.helpers.DefaultHandler;
 
-import javax.xml.parsers.ParserConfigurationException;
 import javax.xml.parsers.SAXParser;
 import javax.xml.parsers.SAXParserFactory;
 import java.io.File;
@@ -25,15 +19,21 @@ import java.lang.reflect.Method;
 import java.net.URL;
 import java.util.*;
 import java.util.jar.JarFile;
+import java.util.regex.Pattern;
 import java.util.zip.ZipEntry;
 
 public class DeployTimeEnhancer {
     private static final Logger LOGGER = Logger.getInstance(LogCategory.OPENEJB_DEPLOY, DeployTimeEnhancer.class);
     private static final SAXParserFactory SAX_PARSER_FACTORY = SAXParserFactory.newInstance();
 
+    private static final String OPENEJB_JAR_ENHANCEMENT_INCLUDE = "openejb.jar.enhancement.include";
+    private static final String OPENEJB_JAR_ENHANCEMENT_EXCLUDE = "openejb.jar.enhancement.exclude";
+
     private static final String CLASS_EXT = ".class";
     private static final String PROPERTIES_FILE_PROP = "propertiesFile";
     private static final String META_INF_PERSISTENCE_XML = "META-INF/persistence.xml";
+    private static final String TMP_ENHANCEMENT_SUFFIX = ".tmp-enhancement";
+    private static final String JAR_ENHANCEMENT_SUFFIX = "-enhanced";
 
     static {
         SAX_PARSER_FACTORY.setNamespaceAware(true);
@@ -53,7 +53,7 @@ public class DeployTimeEnhancer {
             cstr = arg2.getConstructor(Properties.class);
             mtd = enhancerClass.getMethod("run", String[].class, arg2);
         } catch (Exception e) {
-            LOGGER.warning("openjpa enhancer can't be found in the contanier, will be skipped");
+            LOGGER.warning("openjpa enhancer can't be found in the container, will be skipped");
             mtd = null;
             cstr = null;
         }
@@ -82,7 +82,7 @@ public class DeployTimeEnhancer {
                     ZipEntry entry = jar.getEntry(META_INF_PERSISTENCE_XML);
                     if (entry != null) {
                         final String path = file.getAbsolutePath();
-                        final File unpacked = new File(path.substring(0, path.length() - 4));
+                        final File unpacked = new File(path.substring(0, path.length() - 4) + TMP_ENHANCEMENT_SUFFIX);
                         JarExtractor.extract(file, unpacked);
                         feed(classesByPXml, new File(unpacked, META_INF_PERSISTENCE_XML).getAbsolutePath());
                     }
@@ -116,10 +116,20 @@ public class DeployTimeEnhancer {
         // clean up extracted jars
         for (Map.Entry<String, List<String>> entry : classesByPXml.entrySet()) {
             final List<String> values = entry.getValue();
-            for (String path : values) {
-                final File file = new File(path + ".jar");
-                if (file.exists()) {
-                    Files.delete(file);
+            for (String rawPath : values) {
+                if (rawPath.endsWith(TMP_ENHANCEMENT_SUFFIX)) {
+                    final String path = rawPath.substring(0, rawPath.length() - TMP_ENHANCEMENT_SUFFIX.length());
+                    final File dir = new File(path);
+                    final File file = new File(path + ".jar");
+                    if (file.exists()) {
+                        try {
+                            JarCreator.jarDir(dir, new File(dir.getParentFile(), dir.getName() + JAR_ENHANCEMENT_SUFFIX + ".jar"));
+                            Files.delete(file); // don't delete if any exception is thrown
+                        } catch (IOException e) {
+                            LOGGER.error("can't repackage enhanced jar file " + file.getName());
+                        }
+                        Files.delete(dir);
+                    }
                 }
             }
             values.clear();
@@ -190,9 +200,20 @@ public class DeployTimeEnhancer {
     }
 
     private static class ClassFilter implements FileFilter {
+        private static final String DEFAULT_INCLUDE = "\\*";
+        private static final String DEFAULT_EXCLUDE = "";
+        private static final Pattern INCLUDE_PATTERN = Pattern.compile(SystemInstance.get().getOptions().get(OPENEJB_JAR_ENHANCEMENT_INCLUDE, DEFAULT_INCLUDE));
+        private static final Pattern EXCLUDE_PATTERN = Pattern.compile(SystemInstance.get().getOptions().get(OPENEJB_JAR_ENHANCEMENT_EXCLUDE, DEFAULT_EXCLUDE));
+
         @Override
         public boolean accept(final File file) {
-            return file.getName().endsWith(CLASS_EXT);
+            boolean isClass = file.getName().endsWith(CLASS_EXT);
+            if (DEFAULT_EXCLUDE.equals(EXCLUDE_PATTERN) && DEFAULT_INCLUDE.equals(INCLUDE_PATTERN)) {
+                return isClass;
+            }
+
+            final String path = file.getAbsolutePath();
+            return isClass && INCLUDE_PATTERN.matcher(path).matches() && !EXCLUDE_PATTERN.matcher(path).matches();
         }
     }
